requirements.txt
configobj
elasticsearch>=5.0.0,<6.0.0
scandir
tika
watchdog
docopt
persist-queue
appdirs
jsondiff
zerorpc
python-magic-bin
PyPDF2
netCDF4
Pillow
olefile
python-docx
openpyxl
pyyaml
